@@ -1,13 +1,14 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Thayen.WebPlanner.Data
+namespace Thayen.WebPlanner.Data;
+
+public static class ServiceCollectionExtensions
 {
-    public static class ServiceCollectionExtensions
+    public static IServiceCollection AddDataServices(this IServiceCollection collection, string connectionString)
     {
-        public static IServiceCollection AddDataServices(this IServiceCollection collection, string connectionString)
-        {
-            return collection.AddDbContextPool<PlannerDatabaseContext>(options => options.UseNpgsql(connectionString));
-        }
+        return collection.AddDbContextPool<PlannerDatabaseContext>(options =>
+            options.UseNpgsql(connectionString)
+                   .UseSnakeCaseNamingConvention());
     }
 }
